refactor conftest.py files
